create handler file that includes all app methods
